requirements.txt
langchain==0.0.259
pypdf==3.15.0
chromadb==0.4.5
openai~=0.27.8
tiktoken~=0.4.0
typing-inspect==0.9.0
typing_extensions==4.7.1
python-dotenv~=1.0.0
pathlib~=1.0.1
pydantic==2.1.1


pandas~=2.0.3
scipy~=1.11.1
pypdf2~=3.0.1